+++ projects/ion/src/lib/core/types/index.ts
@@ -3,3 +3,4 @@ export * from './info-badge';
 export * from './status';
 export * from './tooltip';
 export * from './card';
+export * from './tab-group';
