+++ projects/ion/src/lib/date-picker/date-picker/date-picker.component.scss
@@ -176,7 +176,7 @@
   & ::ng-deep div.footer {
     display: flex;
     align-items: center;
-    justify-content: end;
+    justify-content: flex-end;
     gap: 28px;
     margin-top: 26px;
   }
