+++ projects/ion/src/lib/tab-group/tab-group.component.ts
@@ -1,18 +1,6 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
-import { BorderDirectionType, DirectionType } from '../core/types';
-import { IonTabProps, TabSize } from '../tab/tab.component';
-
-export interface TabInGroup extends IonTabProps {
-  selected: boolean;
-}
-
-export interface TabGroupProps {
-  tabs: TabInGroup[];
-  border?: BorderDirectionType;
-  direction: DirectionType;
-  size?: TabSize;
-  selected: EventEmitter<TabInGroup>;
-}
+import { BorderDirectionType, DirectionType, TabInGroup } from '../core/types';
+import { TabSize } from '../tab/tab.component';
 
 @Component({
   selector: 'ion-tab-group',
